Colorize also inline snippets
